# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital Recipe Box API",
    "description": "REST API for Digital Recipe Box - A comprehensive recipe management system with Google OAuth authentication",
    "version": "1.0.0",
    "contact": {
      "name": "CSE341 Team-12",
      "url": "https://github.com/CSE341-TEAM-12/digital-recipe-box"
    }
  },
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "googleAuth": {
      "type": "oauth2",
      "authorizationUrl": "/auth/google",
      "flow": "implicit",
      "scopes": {
        "profile": "Access to user profile information",
        "email": "Access to user email address"
      },
      "description": "Google OAuth 2.0 authentication. Use /auth/google to initiate login flow."
    },
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token for authentication. Format: Bearer {token} (For testing purposes only)"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/login/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/auth/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/login/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}